### 2020-05-27 72bc19d
improve location code handling (#180)
* improve location code handling

* flake8

* fix f strings
