### 2020-05-29 c0aa1f8
add get_stations_kwargs option to df_to_inv
