+++ tests/test_utils/test_stations_utils.py
@@ -8,8 +8,10 @@ import pytest
 
 import obsplus
 from obsplus.constants import NSLC, DF_TO_INV_COLUMNS
-from obsplus.utils.stations import df_to_inventory
+from obsplus.utils import get_seed_id_series
 from obsplus.utils.misc import suppress_warnings
+from obsplus.utils.stations import df_to_inventory
+from obsplus.utils.time import to_utc
 
 
 class TestDfToInventory:
@@ -46,45 +48,6 @@ class TestDfToInventory:
         row1 = numerics + ["2019-01-01", "2020-01-01"]
         row2 = numerics + ["2020-01-01", "2200-01-01"]
         return pd.DataFrame([row1, row2], columns=DF_TO_INV_COLUMNS)
-
-    @pytest.fixture
-    def df_with_response(self):
-        """
-        Add response information to the dataframe.
-
-        Note: The datalogger is probably not correct for the test station.
-        It is just to ensure the NRL can be used to get a response.
-        """
-        _inv = obsplus.load_dataset("bingham_test").station_client.get_stations()
-        inv = _inv.select(station="NOQ")
-
-        with suppress_warnings():
-            df = obsplus.stations_to_df(inv)
-
-        # set instrument str
-        sensor_keys = ("Nanometrics", "Trillium 120 Horizon")
-        # get digitizer keys
-        datalogger_keys = (
-            "Nanometrics",
-            "Centaur",
-            "1 Vpp (40)",
-            "Off",
-            "Linear phase",
-            "100",
-        )
-        # keep one as a tuple and convert the other to str
-        df["sensor_keys"] = [sensor_keys for _ in range(len(df))]
-        df["datalogger_keys"] = "__".join(datalogger_keys)
-        # drop seed id
-        df = df.drop(columns="seed_id")
-        return df
-
-    @pytest.fixture
-    def df_with_partial_responses(self, df_with_response):
-        """ test creating inv with partial responses. """
-        # set one row to None
-        df_with_response.loc[0, "sensor_keys"] = None
-        return df_with_response
 
     def test_type(self, inv_from_df):
         """ An inv should have been returned. """
@@ -141,33 +104,6 @@ class TestDfToInventory:
         inv_sub = inv.get_stations(**kwargs)
         assert inv_sub[0][0][0].azimuth is None
 
-    @pytest.mark.requires_network
-    def test_response(self, df_with_response):
-        """ Ensure the NRL is used to pull responses. """
-        with suppress_warnings():
-            inv = df_to_inventory(df_with_response)
-        assert isinstance(inv, obspy.Inventory)
-        for net in inv:
-            for sta in net:
-                for chan in sta:
-                    assert chan.response is not None
-
-    @pytest.mark.requires_network
-    def test_response_one_missing(self, df_with_partial_responses):
-        """ Ensure responses which can be got are fetched. """
-        df = df_with_partial_responses
-        with suppress_warnings():
-            inv = df_to_inventory(df_with_partial_responses)
-
-        missing = df["sensor_keys"].isnull() | df["datalogger_keys"].isnull()
-        missing_chan_data = set(df[missing]["channel"])
-
-        for net in inv:
-            for sta in net:
-                for chan in sta:
-                    name = chan.code
-                    assert chan.response is not None or name in missing_chan_data
-
     def test_make_station_level_inventory(self, df_from_inv):
         """ Ensure station level invs can be constructed. """
         df = df_from_inv.drop(columns="channel")
@@ -190,3 +126,139 @@ class TestDfToInventory:
         inv = df_to_inventory(df)
         for channel in inv.get_contents()["channels"]:
             assert channel.split(".")[2] == "00"
+
+
+@pytest.mark.requires_network
+class TestDfToInventoryGetResponses:
[…]
