### 2020-05-29 55931d9
revamp compose_docstrings
