+++ obsplus/utils/stations.py
@@ -19,13 +19,13 @@ from obsplus.constants import (
     SMALLDT64,
     LARGEDT64,
     NSLC,
-    GET_STATIONS_UTC_KWARGS,
+    STATION_DTYPES,
 )
 from obsplus.exceptions import AmbiguousResponseError
 from obsplus.interfaces import StationClient
-from obsplus.utils.time import to_utc
-from obsplus.constants import STATION_DTYPES
 from obsplus.utils.docs import compose_docstring, format_dtypes
+from obsplus.utils.misc import FunctionCacheDescriptor
+from obsplus.utils.time import to_utc
 
 LARGE_NUMBER = obspy.UTCDateTime("3000-01-01").timestamp
 
@@ -41,69 +41,37 @@ type_mappings = {"start_date": to_utc, "end_date": to_utc}
 station_col_str = format_dtypes(STATION_DTYPES)
 
 
-@compose_docstring(station_columns=station_col_str)
-def df_to_inventory(
-    df: pd.DataFrame, client: Optional[StationClient] = None
-) -> obspy.Inventory:
-    """
-    Create a station inventory from a dataframe.
-
-    Parameters
-    ----------
-    df
-        A dataframe which must have the the following columns:
-
-        {station_columns}
-
-        :func:`obsplus.stations_to_df`.
-    client
-        Any client
+def _make_key_mappings(cls):
+    """ Create a mapping from columns in df to kwargs for cls. """
+    base_params = set(inspect.signature(cls).parameters)
+    new_map = mapping_keys[cls]
+    base_map = {x: x for x in base_params - set(new_map)}
+    base_map.update(new_map)
+    return base_map
 
 
-    Notes
-    -----
-    There are two ways to include response information:
+class _InventoryConstructor:
+    """A private helper class for constructing inventories from dataframes."""
 
-    1. Using the Nominal Response Library (NRL):
-    (https://docs.obspy.org/master/packages/obspy.clients.nrl.html)
-    If the dataframe has columns named "sensor_keys" and "datalogger_keys"
-    these will indicate the response information should
-    be fetched using ObsPy's NRL client. Each of these columns should either
-    contain tuples or strings where the keys are separated by double
-    underscores (__).
-    For example, to specify sensor keys:
-    ('Nanometrics', 'Trillium 120 Horizon')
-    or
-    'Nanometrics__Trillium 120 Horizon'
-    are both valid.
+    _client_col = "get_station_kwargs"
+    nrl_response_cols = frozenset({"datalogger_keys", "sensor_keys"})
+    # get cached expected kwargs for inventory class constructors
+    net_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Network))
+    sta_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Station))
+    cha_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Channel))
+    # columns for performing groupby at various levels
+    _gb_cols = dict(
+        network=("network",),
+        station=("station", "start_date", "end_date"),
+        channel=("channel", "location", "start_date", "end_date"),
+    )
 
-    2. Using a station client:
-    If the dataframe contains a column get_stations_kwargs it indicates that
-    either a client was passed as the client argument or the fdsn IRIS client
-    should be used to download station information. The contents of this column
-    must be a dictionary of acceptable keyword arguments for the client's
-    `get_stations` method. All time values must be provided as iso8601
-    strings.
-    For example,
-    {'network': 'UU', 'station': 'TMU', 'location': '01', 'channel': 'HHZ',
-     'starttime': '2017-01-02',}
-    would use either a provided station client (or ObsPy's default FDSN client
-    if non is provided) to download a response for the corresponding channel.
[…]
