+++ CHANGELOG.txt
@@ -7,12 +7,16 @@ Master
   - obsplus.EventBank
     * Fixed issue where path_structure was overriding a user-specified value
       of `""` (#178)
-  - obsplus.WaveBank
-    * Fixed issue where path_structure was overriding a user-specified value
-      of `""` (#178)
   - obsplus.structures.fetcher
     * fixed issue with Fetcher raising Pandas error when an inventory with
       duplicate chanenls was used (#183, #184).
+  - obsplus.utils
+    * Refactored df_to_inventory (#182).
+    * Refactored compose_docstring and added function for printing dataframe
+      columns and their expected dtypes (#182).
+  - obsplus.WaveBank
+    * Fixed issue where path_structure was overriding a user-specified value
+      of `""` (#178)
 
 obsplus 0.1.0:
   - obsplus.structures.fetcher
