+++ setup.cfg
@@ -23,6 +23,7 @@ extend_ignore =
     D401
     D403
     D404
+    D412
     D414
     W605
 ; exclude certain files
