### 2020-05-30 d3d31d6
fix fetcher duplicate channel issue. (#184)
* fix #183

* reference correct series
