### 2020-05-30 60a9583
improve docstring
