### 2020-06-01 096ebaa
Client option to get response in df_to_inventory (#182)
* add get_stations_kwargs option to df_to_inv

* revamp compose_docstrings
