+++ obsplus/utils/stations.py
@@ -2,8 +2,11 @@
 Utilities for working with inventories.
 """
 import inspect
+
+import json
 from functools import singledispatch, lru_cache
 from pathlib import Path
+from typing import Optional
 
 import numpy as np
 import obspy
@@ -12,8 +15,18 @@ from obspy.core.inventory import Channel, Station, Network
 
 import obsplus
 import obsplus.utils.misc
-from obsplus.constants import station_clientable_type, SMALLDT64, LARGEDT64, NSLC
+from obsplus.constants import (
+    station_clientable_type,
+    SMALLDT64,
+    LARGEDT64,
+    NSLC,
+    STATION_DTYPES,
+)
+from obsplus.exceptions import AmbiguousResponseError
 from obsplus.interfaces import StationClient
+from obsplus.utils import get_nslc_series
+from obsplus.utils.docs import compose_docstring, format_dtypes
+from obsplus.utils.misc import FunctionCacheDescriptor
 from obsplus.utils.time import to_utc
 
 LARGE_NUMBER = obspy.UTCDateTime("3000-01-01").timestamp
@@ -27,37 +40,40 @@ mapping_keys = {
 }
 
 type_mappings = {"start_date": to_utc, "end_date": to_utc}
+station_col_str = format_dtypes(STATION_DTYPES)
 
 
-def df_to_inventory(df) -> obspy.Inventory:
-    """
-    Create a station inventory from a dataframe.
+def _make_key_mappings(cls):
+    """ Create a mapping from columns in df to kwargs for cls. """
+    base_params = set(inspect.signature(cls).parameters)
+    new_map = mapping_keys[cls]
+    base_map = {x: x for x in base_params - set(new_map)}
+    base_map.update(new_map)
+    return base_map
 
-    Parameters
-    ----------
-    df
-        A dataframe which must have the same columns as the once produced by
-        :func:`obsplus.stations_to_df`.
 
-    Notes
-    -----
-    The dataframe can also contain columns named "sensor_keys" and
-    "datalogger_keys" which will indicate the response information should
-    be fetched suing obspy's ability to interact with the nominal response
-    library. Each of these columns should either contain tuples or strings
-    where the keys are separated by double underscores (__).
-    """
+class _InventoryConstructor:
+    """A private helper class for constructing inventories from dataframes."""
 
-    def _make_key_mappings(cls):
-        """ Create a mapping from columns in df to kwargs for cls. """
-        base_params = set(inspect.signature(cls).parameters)
-        new_map = mapping_keys[cls]
-        base_map = {x: x for x in base_params - set(new_map)}
-        base_map.update(new_map)
-        return base_map
+    _client_col = "get_station_kwargs"
+    _nrl_response_cols = frozenset({"datalogger_keys", "sensor_keys"})
+    # get cached expected kwargs for inventory class constructors
+    net_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Network))
+    sta_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Station))
+    cha_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Channel))
+    # columns for performing groupby at various levels
+    _gb_cols = dict(
+        network=("network",),
+        station=("station", "start_date", "end_date"),
+        channel=("channel", "location", "start_date", "end_date"),
+    )
 
-    def _groupby_if_exists(df, columns):
+    def __init__(self, station_client=None):
+        self._client = station_client
+
+    def _groupby_if_exists(self, df, level):
         """ Groupby columns if they exist on dataframe, else return empty. """
+        columns = list(self._gb_cols[level])
         cols = list(obsplus.utils.misc.iterate(columns))
[…]

[…]
