### 2020-06-01 6f719b5
quickfix for df_to_inv
