+++ CHANGELOG.txt
@@ -1,4 +1,4 @@
-Master
+obsplus 0.1.1
   - obsplus.events
     * Fixed issue with validators flagging picks on stations with the same
       name but different networks (#173).
