### 2020-06-09 #2236 @leouieda
Hi @arfon I'm happy to edit this paper :+1:

@d-chambers I'll be a bit more free next week and will start looking over the paper then. Thank you for your patience :slightly_smiling_face:
