+++ tests/test_stations/test_get_stations.py
@@ -33,7 +33,7 @@ class TestGetStation:
             station="BOB",
             channel="HHZ",
             start_date="2019-01-01",
-            enddate="2100-01-01",
+            end_date="2100-01-01",
             sample_rate=250,
             latitude=0,
             longitude=0,
@@ -46,7 +46,7 @@ class TestGetStation:
             station="01",
             channel="BHZ",
             start_date="2019-01-01",
-            enddate="2100-01-01",
+            end_date="2100-01-01",
             sample_rate=1000,
             latitude=0,
             longitude=0,
