### 2020-06-16 af0837f
fix duplicate channel df to inv issue
