### 2020-06-16 d8c0784
fix coverage issue, correct issues in changelog
