### 2020-06-16 4e935f2
fix duplicate channel issue on df_to_inventory (#190)
