+++ obsplus/utils/stations.py
@@ -1,9 +1,8 @@
 """
 Utilities for working with inventories.
 """
-import inspect
-
 import json
+import warnings
 from functools import singledispatch, lru_cache
 from pathlib import Path
 from typing import Optional
@@ -19,6 +18,8 @@ from obsplus.constants import (
     station_clientable_type,
     SMALLDT64,
     LARGEDT64,
+    SMALL_UTC,
+    BIG_UTC,
     NSLC,
     STATION_DTYPES,
 )
@@ -26,7 +27,6 @@ from obsplus.exceptions import AmbiguousResponseError
 from obsplus.interfaces import StationClient
 from obsplus.utils import get_nslc_series
 from obsplus.utils.docs import compose_docstring, format_dtypes
-from obsplus.utils.misc import FunctionCacheDescriptor
 from obsplus.utils.time import to_utc
 
 LARGE_NUMBER = obspy.UTCDateTime("3000-01-01").timestamp
@@ -43,28 +43,51 @@ type_mappings = {"start_date": to_utc, "end_date": to_utc}
 station_col_str = format_dtypes(STATION_DTYPES)
 
 
-def _make_key_mappings(cls):
-    """ Create a mapping from columns in df to kwargs for cls. """
-    base_params = set(inspect.signature(cls).parameters)
-    new_map = mapping_keys[cls]
-    base_map = {x: x for x in base_params - set(new_map)}
-    base_map.update(new_map)
-    return base_map
-
-
 class _InventoryConstructor:
     """A private helper class for constructing inventories from dataframes."""
 
     _client_col = "get_station_kwargs"
     _nrl_response_cols = frozenset({"datalogger_keys", "sensor_keys"})
-    # get cached expected kwargs for inventory class constructors
-    net_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Network))
-    sta_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Station))
-    cha_map = FunctionCacheDescriptor(lambda: _make_key_mappings(Channel))
+    # define expected kwargs for inventory class constructors
+    net_map = {
+        "code": "network",
+        "operator": "operator",
+    }
+    sta_map = {
+        "code": "station",
+        "equipment": "equipment",
+        "creation_date": "creation_date",
+        "latitude": "latitude",
+        "longitude": "longitude",
+        "elevation": "elevation",
+    }
+    cha_map = {
+        "response": "response",
+        "start_date": "start_date",
+        "end_date": "end_date",
+        "code": "channel",
+        "location_code": "location",
+        "sample_rate": "sample_rate",
+        "sensor": "sensor",
+        "equipment": "equipment",
+        "elevation": "elevation",
+        "azimuth": "azimuth",
+        "dip": "dip",
+        "data_logger": "data_logger",
+        "latitude": "latitude",
+        "longitude": "longitude",
+        "depth": "depth",
+    }
+    # A set of expected columns
+    _expected_cols = (
+        set(net_map.values()) | set(cha_map.values()) | set(sta_map.values())
+    )
+    # A set of columns to drop before building inventory
+    _drop_cols = list(_nrl_response_cols) + [_client_col, "index", "seed_id"]
     # columns for performing groupby at various levels
     _gb_cols = dict(
         network=("network",),
-        station=("station", "start_date", "end_date"),
+        station=("station", "latitude", "longitude", "elevation"),
         channel=("channel", "location", "start_date", "end_date"),
     )
 
@@ -153,8 +176,6 @@ class _InventoryConstructor:
     def _update_nrl_response(self, response, df):
         """Update the responses with NRL."""
         # df doesn't have needed columns, just exit
[…]
