### 2020-06-26 ad696d8
Broaden the default columns for the dataframe extractors
