### 2020-06-27 5562e8e
Tweak tests to work with NRL change
