+++ CHANGELOG.txt
@@ -8,6 +8,9 @@ obsplus master
   - obsplus.EventBank
     * Added `overwrite_existing` keyword to `put_events` to disable squashing
       existing events if desired (#191).
+  - obsplus.events.pd
+    * Made sure that the pre-defined extractors (`*_to_df`) still return all of
+      the expected columns if there are no valid objects.
 
 obsplus 0.1.1
   - obsplus.events
