### 2020-06-27 ffa6abc
Broaden the default columns for the dataframe extractors (#192)
* Broaden the default columns for the dataframe extractors

* Black issues

* Tweak tests to work with NRL change
