### 2020-07-14 dfcbb6d
Fix picks_to_df with location code with decimal (#193)
