### 2020-07-24 #2236 @leouieda
@d-chambers apologies for the delay in getting this started. I've been going over the paper and code and have a few comments. These are mostly issues you could start working on in the mean time which are very likely to be raised during review:

* Add a citation to pandas in the paper
* Link to the contributing guide on the README
* Add instructions for running your tests (reviewers will likely want to do this)
* I found example usage in the [Introduction](https://niosh-mining.github.io/obsplus/versions/latest/notebooks/intro.html) page but that took a bit of digging since it wasn't clear from the heading and link texts that these were examples, not API reference documentation.

I'll start inviting reviewers but this will give you a bit of a head start.
