### 2020-07-25 2f07795
small pre-commit fixes
