+++ docs/conf.py
@@ -79,9 +79,6 @@ numpydoc_show_class_members = False
 # The master toctree document.
 master_doc = "index"
 
-# General information about the project.
-project = "ObsPlus"
-copyright = f"{year}, Derrick Chambers"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -91,6 +88,10 @@ copyright = f"{year}, Derrick Chambers"
 version = obsplus.__version__
 # The full version, including alpha/beta/rc tags.
 release = obsplus.__version__
+
+# General information about the project.
+project = f"ObsPlus  [ {version} ]"
+copyright = f"{year}, Derrick Chambers"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
