### 2020-07-27 3ae8213
leouieda pre review (#194)
address issues raised by leouieda for JOSS review
