+++ .github/workflows/on_master_commits.yml
@@ -5,6 +5,7 @@ on:
       - master
 
 jobs:
+  # Calculates new coverage for the base branch
   calc_coverage:
     runs-on: ubuntu-latest
 
@@ -47,3 +48,44 @@ jobs:
           name: codecov-umbrella
           yml: ./codecov.yml
           fail_ci_if_error: true
+
+
+  update_gh_pages:
+    # Build new documentation for latest
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v1
+
+      - name: Setup conda
+        uses: goanpeca/setup-miniconda@v1
+        with:
+          miniconda-version: 'latest'
+          python-version: 3.7
+          activate-environment: test
+          environment-file: .github/test_conda_env.yml
+          condarc-file: .github/test_condarc.yml
+
+      - name: install obsplus
+        shell: bash -l {0}
+        run: |
+          pip install -e .[dev]
+
+      - name: git config
+        shell: bash -l {0}
+        run: |
+          git config --global user.email "donotemailme@example.com"
+          git config --global user.name "GithubActions"
+
+      - name: build docs
+        shell: bash -l {0}
+        run: |
+          python scripts/stage_docs.py --remove-dirty
+
+      - name: push
+        uses: ad-m/github-push-action@master
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          branch: gh-pages
+          force: true
+          directory: ../obsplus_documentation
