### 2020-07-28 #2236 @megies
@leouieda I'm actually not using pandas a lot, so maybe there could be better options. Maybe @thomaslecocq? But if no other people come forward I can be a backup.
