### 2020-07-28 #2236 @leouieda
@ThomasLecocq would be willing to review this submission for JOSS?
