### 2020-07-28 #2236 @anowacki
@leouieda Apologies, I'm afraid I don't think I would be useful here, so I would be grateful if you were able to find another reviewer.

(I should update the list of things on the reviewers spreadsheet to reflect the fact that I don't  really use Python any more, have never used Pandas, and so can't give much in the way of a helpful review.)
