### 2020-07-31 5878f89
Meaningful error message if required input cols missing for SpatialCalculator
