+++ CHANGELOG.txt
@@ -8,6 +8,8 @@ obsplus master
     * Fixed an issue with `picks_to_df` which would return the wrong channel
       code if the location code had a decimal. Now a ValueError is raised if
       any seed ids are found which don't have exactly 3 decimals (see #193).
+    * Made sure that `SpatialCalculator` returns a meaningful error message if
+      required columns are missing from DataFrame inputs (#197).
   - obsplus.EventBank
     * Added `overwrite_existing` keyword to `put_events` to disable squashing
       existing events if desired (#191).
