### 2020-07-31 e43aa80
Dynamically created dataset (#196)
* change behavior slightly

* fix datetime casting

* changelog
