### 2020-07-31 e94fb5b
Added description of new tests
