### 2020-08-01 2e3d1f2
SpatialCalculator DataFrame input columns (#197)
* Meaningful error message if required input cols missing for SpatialCalculator

* Update CHANGELOG, blacken

* Re-blackened...

* Argh!

* Fixed the stuff that I broke

* Merge and blacken

* Boost test coverage

* Added description of new tests
