### 2020-09-11 #2236 @lheagy
Hi @leouieda, thanks for thinking of me -- this looks like an interesting project! I am pretty loaded at the moment unfortunately, so I don't have the bandwidth to take this on. I wonder if, given the focus on the #subsurface channel of Swung on interoperability between software packages / data representations (xarray, pandas, etc), if there would be someone there who would be interested?
