### 2020-09-12 2711de0
Standardize location code behavior
