### 2020-09-12 c6de896
Standardize location code behavior
