+++ tests/test_stations/test_pd_stations.py
@@ -60,7 +60,9 @@ class TestInv2Df:
         assert invdf["start_date"].dt  # if not dt this will raise
         assert invdf["end_date"].dt
 
-    @pytest.mark.parametrize("input,expected", [(1, "01"), (1.0, "01"), ("01", "01"), ("1", "1")])
+    @pytest.mark.parametrize(
+        "input,expected", [(1, "01"), (1.0, "01"), ("01", "01"), ("1", "1")]
+    )
     def test_location_codes(self, invdf, input, expected):
         """ make sure location codes are handled nicely """
         invdf = invdf.copy()
@@ -77,7 +79,9 @@ class TestInv2Df:
         invdf = stations_to_df(invdf)
         assert (invdf.loc[invdf["station"] == "RJOB"]["location"] == expected).all()
         assert (invdf.loc[invdf["station"] != "RJOB"]["location"] == "").all()
-        assert (invdf["location"] == invdf["seed_id"].str.split(".", expand=True)[2]).all()  # This one will be a little bit tougher
+        assert (
+            invdf["location"] == invdf["seed_id"].str.split(".", expand=True)[2]
+        ).all()  # This one will be a little bit tougher
 
 
 class TestReadInventory:
