### 2020-09-12 b0f178c
Minor cleanup
