### 2020-09-17 95a47fe
EventBank.get_events to ignore deleted files (#198)
* fix deleted files

* changelog

* reblacken

* remove redundant linting installs

* remove exists requirement
