### 2020-09-22 ac2e1d5
NSLC cleanup (#199)
* Standardize location code behavior

* Standardize location code behavior

* Minor cleanup

* Trying to deal with lint issue?

* Update changelog
