+++ obsplus/constants.py
@@ -49,10 +49,10 @@ _DATETIME_TYPE_MAP = {"datetime64[ns]": np.int64, "timedelta64[ns]": np.int64}
 
 # columns required for station data
 STATION_DTYPES = OrderedDict(
-    network=str,
-    station=str,
-    location=str,
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     seed_id=str,
     latitude=float,
     longitude=float,
@@ -68,10 +68,10 @@ STATION_DTYPES = OrderedDict(
 STATION_COLUMNS = tuple(STATION_DTYPES)
 
 DF_TO_INV_DTYPES = OrderedDict(
-    network=str,
-    station=str,
-    location="location_code",
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     latitude=float,
     longitude=float,
     elevation=float,
@@ -145,10 +145,10 @@ PICK_DTYPES = OrderedDict(
     author=str,
     agency_id=str,
     event_id=str,
-    network=str,
-    station=str,
-    location="location_code",
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     uncertainty=float,
     lower_uncertainty=float,
     upper_uncertainty=float,
@@ -201,10 +201,10 @@ AMPLITUDE_DTYPES = OrderedDict(
     agency_id=str,
     event_time="datetime64[ns]",
     event_id=str,
-    network=str,
-    station=str,
-    location="location_code",
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     uncertainty=float,
     lower_uncertainty=float,
     upper_uncertainty=float,
@@ -228,10 +228,10 @@ STATION_MAGNITUDE_DTYPES = OrderedDict(
     agency_id=str,
     event_id=str,
     event_time="datetime64[ns]",
-    network=str,
-    station=str,
-    location="location_code",
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     uncertainty=float,
     lower_uncertainty=float,
     upper_uncertainty=float,
@@ -288,10 +288,10 @@ ARRIVAL_DTYPES = OrderedDict(
     creation_time="datetime64[ns]",
     author=str,
     agency_id=str,
-    network=str,
-    station=str,
-    location="location_code",
-    channel=str,
+    network="nslc_code",
+    station="nslc_code",
+    location="nslc_code",
+    channel="nslc_code",
     origin_id=str,
     origin_time="datetime64[ns]",
 )
