### 2020-09-23 18911c6
fix dtype address review
