### 2020-09-23 8ad3eab
associate_merge (#200)
* work on tests

* add merge_and tests

* changelog

* fix dtype address review
