### 2020-09-24 4af8924
Fix citation to Pandas in the JOSS paper
The citation key being used didn't match the one in `paper.bib`.
