### 2020-09-24 #2696 @whedon
[REVIEW]: ObsPlus: A Pandas Centric ObsPy Expansion Pack
**Submitting author:** @d-chambers (<a href="http://orcid.org/0000-0003-3656-6607">Derrick Chambers</a>)
**Repository:** <a href="https://github.com/niosh-mining/obsplus" target ="_blank">https://github.com/niosh-mining/obsplus</a>
**Version:** v0.1.0
