### 2020-09-26 941081e
make update_timestamp an attribute
