+++ tests/test_bank/test_eventbank.py
@@ -666,16 +666,18 @@ class TestPutEvents:
         mtimes_2 = [x.stat().st_mtime for x in files]
         assert mtimes_1 == mtimes_2
 
-    def test_put_events_no_update_index_timestamp(self, ebank):
+    def test_put_events_no_update_index_timestamp(self, ebank, monkeypatch):
         """
         Ensure events can be put into the bank without updating the index
         timestamp.
         """
+        monkeypatch.setattr(ebank, "allow_update_timestamp", False)
         last_update = ebank.last_updated
         cat = obspy.read_events()
-        ebank.put_events(cat, update_timestamp=False)
+        ebank.put_events(cat)
         time.sleep(0.001)
         new_update_time = ebank.last_updated
+        # The last updated timestamp should not have changed.
         assert last_update == new_update_time
 
 
