### 2020-09-26 bd2392c
add update_index argument (#202)
* add update_index argument

* make update_timestamp an attribute
