+++ paper/paper.bib
@@ -1,20 +1,21 @@
 @article{Krischer:2015,
-  	Author = {{Krischer}, L., {Megies}, T., {Barsch}, R., {Beyreuther}, M., {Lecocq}, T., {Caudron}, C. and {Wassermann}, J.,},
-  	Journal = {Computational Science & Discovery},
+  	Author   = {{Krischer}, L., {Megies}, T., {Barsch}, R., {Beyreuther}, M., {Lecocq}, T., {Caudron}, C. and {Wassermann}, J.,},
+  	Journal  = {Computational Science & Discovery},
   	Keywords = {seismology},
-  	Title = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
-  	Year = 2015
+  	Title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
+    doi      = { 10.1088/1749-4699/8/1/014003 }
+  	Year     = 2015
 }
 
 
 @article{Schorlemmer:2011,
-  	Author = {{Schorlemmer}, D., {Euchner}, F., {Kästli}, P. and {Saul}, J.},
-  	Eprint = {1703.04627},
-  	Journal = {Annals of Geophysics},
+  	Author   = {{Schorlemmer}, D., {Euchner}, F., {Kästli}, P. and {Saul}, J.},
+  	Eprint   = {1703.04627},
+  	Journal  = {Annals of Geophysics},
   	Keywords = {seismology},
-  	Month = mar,
-  	Title = {{QuakeML: status of the XML-based seismological data exchange format}},
-  	Year = 2011
+  	Month    = mar,
+  	Title    = {{QuakeML: status of the XML-based seismological data exchange format}},
+  	Year     = 2011
 }
 
 
