### 2020-10-14 0df16b6
add 3.8 and 3.9 to test matrix
