+++ .github/workflows/runtests.yml
@@ -25,8 +25,11 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.6, 3.7, 3.8]
+#        os: [ubuntu-latest, macos-latest, windows-latest]
+#        python-version: [3.6, 3.7, 3.8]
+        os: [windows-latest]
+        python-version: [ 3.8 ]
+
 
     steps:
       - uses: actions/checkout@v1
@@ -44,6 +47,11 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install -e .[dev]
+
+      - name: fix pywin
+        if: matix.os == "windows-latest"
+        shell: bash -l {0}
+        run: pip install pywin32==225
 
       - name: print package info
         shell: bash -l {0}
