+++ .github/workflows/runtests.yml
@@ -28,7 +28,7 @@ jobs:
 #        os: [ubuntu-latest, macos-latest, windows-latest]
 #        python-version: [3.6, 3.7, 3.8]
         os: [windows-latest]
-        python-version: [ 3.8 ]
+        python-version: [ 3.7, 3.8 ]
 
 
     steps:
@@ -49,7 +49,7 @@ jobs:
           pip install -e .[dev]
 
       - name: fix pywin
-        if: matix.os == "windows-latest"
+        if: matrix.os == "windows-latest"
         shell: bash -l {0}
         run: pip install pywin32==225
 
