+++ .github/workflows/runtests.yml
@@ -49,7 +49,7 @@ jobs:
           pip install -e .[dev]
 
       - name: fix pywin
-        if: matrix.os == "windows-latest"
+        if: matrix.os == 'windows-latest'
         shell: bash -l {0}
         run: pip install pywin32==225
 
