+++ .github/workflows/runtests.yml
@@ -25,10 +25,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-#        os: [ubuntu-latest, macos-latest, windows-latest]
-#        python-version: [3.6, 3.7, 3.8]
-        os: [windows-latest]
-        python-version: [ 3.7, 3.8 ]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: [3.6, 3.7, 3.8]
 
 
     steps:
