### 2020-10-21 5a158f6
add 3.8 and 3.9 to test matrix (#205)
