+++ .github/workflows/runtests.yml
@@ -26,7 +26,8 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.6, 3.7]
+        python-version: [3.6, 3.7, 3.8]
+
 
     steps:
       - uses: actions/checkout@v1
@@ -44,6 +45,11 @@ jobs:
         shell: bash -l {0}
         run: |
           pip install -e .[dev]
+
+      - name: fix pywin
+        if: matrix.os == 'windows-latest'
+        shell: bash -l {0}
+        run: pip install pywin32==225
 
       - name: print package info
         shell: bash -l {0}
