### 2020-11-13 #2696 @leouieda
@seisman thanks for the update. I realize that the earlier than usual AGU deadlines are causing problems. Good luck on the job interviews and let me know when you have some more time.

@d-chambers apologies for the unusually long delays with this paper. I really appreciate your patience in these strange times we live in 😞
