+++ .github/workflows/runtests.yml
@@ -33,7 +33,7 @@ jobs:
       - uses: actions/checkout@v1
 
       - name: Setup conda
-        uses: goanpeca/setup-miniconda@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: 'latest'
           python-version: ${{ matrix.python-version }}
