### 2020-11-19 e8fd34c
update conda gh (#208)
