+++ .github/workflows/on_master_commits.yml
@@ -13,7 +13,7 @@ jobs:
       - uses: actions/checkout@v1
 
       - name: Setup conda
-        uses: goanpeca/setup-miniconda@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: 'latest'
           python-version: 3.7
@@ -58,7 +58,7 @@ jobs:
       - uses: actions/checkout@v1
 
       - name: Setup conda
-        uses: goanpeca/setup-miniconda@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: 'latest'
           python-version: 3.7
