+++ .github/workflows/release_published.yml
@@ -12,7 +12,7 @@ jobs:
       - uses: actions/checkout@v1
 
       - name: Setup conda
-        uses: goanpeca/setup-miniconda@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: 'latest'
           python-version: 3.7
@@ -51,7 +51,7 @@ jobs:
       - uses: actions/checkout@v1
 
       - name: Setup conda
-        uses: goanpeca/setup-miniconda@v1
+        uses: conda-incubator/setup-miniconda@v2
         with:
           miniconda-version: 'latest'
           python-version: 3.7
