### 2020-11-26 999f64c
Reject old data rather than delete it on event merging
