### 2020-11-26 56c674e
Cleanup and improve coverage
