### 2020-11-30 77d1db5
Reject old data rather than delete it on event merging (#209)
* Reject old data rather than delete it on event merging

* Cleanup and improve coverage
