+++ CHANGELOG.txt
@@ -15,6 +15,8 @@ obsplus master
       required columns are missing from DataFrame inputs (#197).
     * Fixed issue in df_to_inventory where a location code of '01' was getting
       converted to '1'. (#199)
+    * Changed `merge_events` and `associate merge` to have option to reject old
+      picks/amplitudes, removed option to delete old picks/amplitudes. (#209)
   - obsplus.EventBank
     * Added `overwrite_existing` keyword to `put_events` to disable squashing
       existing events if desired (#191).
