### 2020-12-06 niosh-mining/obsplus#212 @seisman
Missing community guidelines about how to report issues and seek support
The JOSS requires that software give clear community guidelines.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

obsplus has done a good job about "Contributing guidelines", but in both the repository and the documentation, I don't see any instructions about "How to report issues" and "How to seek report" (maybe I just missed them). For users who are not familiar with GitHub, they may not know what to do to seek support.

xref: https://github.com/openjournals/joss-reviews/issues/2696
