### 2020-12-07 niosh-mining/obsplus#210 @d-chambers
Hi @seisman,

There might be multiple issues going on here but there looks to be at least one issue with the test datasets. Can you try deleting or renaming the opsdata folder then running the tests again? It should be located at ~/opsdata

I have seen something similar before when 1) I upgraded from a very old version of ObsPlus to a more recent version and 2) when the dataset download failed before it was completed. Did you have an older version of ObsPlus installed on your laptop or is there any reason the internet connection was interrupted the first time you ran the test suite?
