### 2020-12-07 niosh-mining/obsplus#210 @seisman
After deleting the data in `~/opsdata`, the tests now pass. Thanks.
