### 2020-12-08 niosh-mining/obsplus#212 @d-chambers
Hi,

Good point. I guess I assumed a base-level familiarity with github but that may not be a fair assumption. I will add more explicit instructions.
