+++ docs/index.rst
@@ -13,6 +13,8 @@ ObsPlus
 
 `Contributing <notebooks/contributing.ipynb>`_
 
+`Support Bug Reporting <notebooks/support.ipynb>`_
+
 `Quick Reference <quickref/index.rst>`_
 
 `Full API Documentation <api/obsplus.rst>`_
