### 2020-12-11 40e87f4
Bank raise on unsupported kwargs (#214)
* fix #207

* boltz review
