+++ .github/workflows/runtests.yml
@@ -25,8 +25,10 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.6, 3.7, 3.8]
+#        os: [ubuntu-latest, macos-latest, windows-latest]
+#        python-version: [3.6, 3.7, 3.8]
+        os: [ ubuntu-latest ]
+        python-version: [ 3.7 ]
 
 
     steps:
@@ -64,6 +66,10 @@ jobs:
             git clone https://github.com/d-chambers/opsdata_test
             mv opsdata_test ~/opsdata
           fi
+
+
+      - name: Setup tmate session
+        uses: mxschmitt/action-tmate@v3
 
       # Runs test suite and calculates coverage
       - name: run test suite
