+++ obsplus/events/schema.py
@@ -1,0 +1,612 @@
+"""
+Pydantic schema for ObsPlus event model.
+
+ObsPlus Event Model is a superset of, and compatible with, ObsPy's Event
+model.
+"""
+from datetime import datetime
+from typing import Optional, List
+from uuid import uuid4
+
+import obspy.core.event as ev
+from obsplus.constants import NSLC
+from pydantic import BaseModel, validator, root_validator
+from typing_extensions import Literal
+
+# ----- Type Literals (enum like)
+
+data_used_wave_type = Literal[
+    "P waves", "body waves", "surface waves", "mantle waves", "combined", "unknown"
+]
+
+AmplitudeCategory = Literal["point", "mean", "duration", "period", "integral", "other"]
+
+AmplitudeUnit = Literal["m", "s", "m/s", "m/(s*s)", "m*s", "dimensionless", "other"]
+
+DataUsedWaveType = Literal[
+    "P waves", "body waves", "surface waves", "mantle waves", "combined", "unknown"
+]
+
+EvaluationMode = Literal["manual", "automatic"]
+
+EvaluationStatus = Literal["preliminary", "confirmed", "reviewed", "final", "rejected"]
+
+EventDescriptionType = Literal[
+    "felt report",
+    "Flinn-Engdahl region",
+    "local time",
+    "tectonic summary",
+    "nearest cities",
+    "earthquake name",
+    "region name",
+]
+
+EventType = Literal[
+    "not existing",
+    "not reported",
+    "earthquake",
+    "anthropogenic event",
+    "collapse",
+    "cavity collapse",
+    "mine collapse",
+    "building collapse",
+    "explosion",
+    "accidental explosion",
+    "chemical explosion",
+    "controlled explosion",
+    "experimental explosion",
+    "industrial explosion",
+    "mining explosion",
+    "quarry blast",
+    "road cut",
+    "blasting levee",
+    "nuclear explosion",
+    "induced or triggered event",
+    "rock burst",
+    "reservoir loading",
+    "fluid injection",
+    "fluid extraction",
+    "crash",
+    "plane crash",
+    "train crash",
+    "boat crash",
+    "other event",
+    "atmospheric event",
+    "sonic boom",
+    "sonic blast",
+    "acoustic noise",
+    "thunder",
+    "avalanche",
+    "snow avalanche",
+    "debris avalanche",
+    "hydroacoustic event",
+    "ice quake",
+    "slide",
+    "landslide",
+    "rockslide",
+    "meteorite",
+    "volcanic eruption",
+]
+
+EventTypeCertainty = Literal["known", "suspected"]
+
+MTInversionType = Literal["general", "zero trace", "double couple"]
+
+MomentTensorCategory = Literal["teleseismic", "regional"]
+
+OriginDepthType = Literal[
+    "from location",
+    "from moment tensor inversion",
+    "from modeling of broad-band P waveforms",
+    "constrained by depth phases",
+    "constrained by direct phases",
+    "constrained by depth and direct phases",
+    "operator assigned",
+    "other",
+]
+
+OriginType = Literal[
+    "hypocenter",
+    "centroid",
+    "amplitude",
+    "macroseismic",
+    "rupture start",
+    "rupture end",
+]
+
+OriginUncertaintyDescription = Literal[
+    "horizontal uncertainty", "uncertainty ellipse", "confidence ellipsoid"
+]
+
+PickOnset = Literal["emergent", "impulsive", "questionable"]
+
+PickPolarity = Literal["positive", "negative", "undecidable"]
+
+SourceTimeFunctionType = Literal["box car", "triangle", "trapezoid", "unknown"]
+
+
+# ----- Type Models
+
+
+class _ObsPyModel(BaseModel):
+    # extra: Optional[Dict[str, Any]] = None
+
+    class Config:
[…]
