+++ requirements.txt
@@ -1,6 +1,8 @@
 numpy >= 1.13.0
-pandas >= 0.24.0
 obspy >= 1.2.1
+pandas >= 0.24.0
 progressbar2
+pydantic
 scipy
 tables
+typing-extensions
