### 2021-01-22 c290876
address calum review, use dict, changelog
