+++ obsplus/events/json.py
@@ -3,7 +3,7 @@ Module for converting obspy object to json format
 """
 
 import json
-from typing import Union
+from typing import Union, Iterable
 
 import obspy
 from obspy.core.event import Event, Catalog
@@ -16,27 +16,37 @@ JSON_SERIALIZER_VERSION = "0.0.0"  # increment when serialization changes
 # -------------------- events to json functions
 
 
-def cat_to_json(events: Union[Catalog, Event]) -> str:
+def _events_to_model(
+    catalog: Union[Catalog, Event, Iterable[Event]]
+) -> event_schema.Catalog:
+    """
+    Convert ObsPy events to Pydantic models.
+    """
+    if isinstance(catalog, Event):  # A single event was passed
+        catalog = Catalog(events=[catalog])
+    elif not isinstance(catalog, Catalog):  # sequence was passed
+        catalog = Catalog(events=catalog)
+    model = event_schema.Catalog.from_orm(catalog)
+    return model
+
+
+def cat_to_json(catalog: Union[Catalog, Event, Iterable[Event]]) -> str:
     """
     Convert events (catalog or event) to json string.
     """
-    if isinstance(events, Event):  # A single event was passed
-        events = Catalog(events=[events])
-    elif not isinstance(events, Catalog):  # sequence was passed
-        events = Catalog(events=events)
-    json_str = event_schema.Catalog.from_orm(events).json()
-    return json_str
+    model = _events_to_model(catalog)
+    return model.json()
 
 
-def cat_to_dict(events: Union[Catalog, Event]) -> dict:
+def cat_to_dict(catalog: Union[Catalog, Event, Iterable[Event]]) -> dict:
     """
     Convert an event object to a
     """
-    json_str = cat_to_json(events)
-    return json.loads(json_str)
+    model = _events_to_model(catalog)
+    return model.dict()
 
 
-def dict_to_cat(cjson) -> Catalog:
+def dict_to_cat(cjson: Union[dict, str]) -> Catalog:
     """Convert a dictionary to a catalog"""
     if isinstance(cjson, str):
         cjson = json.loads(cjson)
