+++ obsplus/utils/misc.py
@@ -115,8 +115,9 @@ def yield_obj_parent_attr(
     >>> assert len(picks)
 
     >>> # --- yield all objects which have resource identifiers
+import obsplus.utils.model
     >>> objects = []  # list of (rid, parent)
-    >>> RID = ev.ResourceIdentifier
+    >>> RID = obsplus.utils.model.ResourceIdentifier
     >>> for rid, parent, attr in yield_obj_parent_attr(cat, cls=RID):
     ...     objects.append((str(rid), parent))
     >>> assert len(objects)
