+++ obsplus/events/schema.py
@@ -6,12 +6,13 @@ model.
 """
 from datetime import datetime
 from typing import Optional, List
-from uuid import uuid4
 
-import obspy.core.event as ev
 from obsplus.constants import NSLC
-from pydantic import BaseModel, validator, root_validator
+from pydantic import root_validator
 from typing_extensions import Literal
+
+
+from obsplus.utils.model import ObsPlusModel, ResourceIdentifier, _ModelWithResourceID
 
 # ----- Type Literals (enum like)
 
@@ -124,74 +125,34 @@ PickPolarity = Literal["positive", "negative", "undecidable"]
 
 SourceTimeFunctionType = Literal["box car", "triangle", "trapezoid", "unknown"]
 
+# --- Subclass of resource_id to indicated referred object type
+# TODO this is a bit of a hack, consider how to improve
+
+
+class ResourceIdentifierPick(ResourceIdentifier):
+    """ID which points to picks."""
+
+
+class ResourceIdentifierOrigin(ResourceIdentifier):
+    """ID which points to origins."""
+
+
+class ResourceIdentifierMagnitude(ResourceIdentifier):
+    """ID which points to magnitudes."""
+
+
+class ResourceIdentifierStationMagnitude(ResourceIdentifier):
+    """ID which points to station magnitudes."""
+
+
+class ResourceIdentifierFocalMechanism(ResourceIdentifier):
+    """ID which points to focal mechanisms."""
+
 
 # ----- Type Models
 
 
-class _ObsPyModel(BaseModel):
-    # extra: Optional[Dict[str, Any]] = None
-
-    class Config:
-        pass
-        validate_assignment = True
-        arbitrary_types_allowed = True
-        orm_mode = True
-        extra = "allow"
-
-    @staticmethod
-    def _convert_to_obspy(value):
-        """Convert an object to obspy or return value."""
-        if hasattr(value, "to_obspy"):
-            return value.to_obspy()
-        return value
-
-    def to_obspy(self):
-        """Convert to obspy objects."""
-        name = self.__class__.__name__
-        cls = getattr(ev, name)
-        out = {}
-        # get schema and properties
-        schema = self.schema()
-        props = schema["properties"]
-        array_props = {x for x, y in props.items() if y.get("type") == "array"}
-        # iterate each property and convert back to obspy
-        for prop in props:
-            val = getattr(self, prop)
-            if prop in array_props:
-                out[prop] = [self._convert_to_obspy(x) for x in val]
-            else:
-                out[prop] = self._convert_to_obspy(val)
-        return cls(**out)
-
-
-class ResourceIdentifier(_ObsPyModel):
-    """Resource ID"""
-
-    id: Optional[str] = None
-
-    @root_validator(pre=True)
-    def get_id(cls, values):
-        """Get the id string from the resource id"""
-        value = values.get("id")
-        if value is None:
-            value = str(uuid4())
-        return {"id": value}
-
-
-class _ModelWithResourceID(_ObsPyModel):
-    """A model which has a resource ID"""
-
-    resource_id: Optional[ResourceIdentifier]
-
-    @validator("resource_id", always=True)
-    def get_resource_id(cls, value):
-        """Ensure a valid str is returned."""
-        if value is None:
-            return str(uuid4())
-        return value
-
[…]
