+++ obsplus/utils/model.py
@@ -1,0 +1,349 @@
+"""
+Logic for defining tree-like data structures.
+"""
+
+from functools import lru_cache
+from typing import Optional, Union, Tuple
+from uuid import uuid4
+
+from obspy.core import event as ev
+from pydantic import root_validator, validator
+from pydantic.main import BaseModel, ModelMetaclass
+
+from obsplus.exceptions import InvalidModelOperation, InvalidModelAttribute
+
+
+# a global cache for graphs
+
+#
+class ObsPlusMeta(ModelMetaclass):
+    """A mixing for the metaclass to add getitem."""
+
+    # def __getitem__(cls, item):
+    #     item_name = str(getattr(item, '__name__', item))
+    #     cls_name = cls.__name__
+    #     name = f"{cls_name}__{item_name}"
+    #
+    #     _dict = {
+    #         '__reduce__': lambda cls: f"{cls.__module__}.{cls.__name__}"
+    #         # '__getstate__': ModelMetaclass.__getstate__,
+    #         # "__setstate__": ModelMetaclass.__setstate__,
+    #     }
+    #     new = type(name, (cls,), _dict)
+    #
+    #
+    #     return new
+    #
+    # def __getstate__(cls):
+    #     breakpoint()
+    #
+    # def __setstate__(cls):
+    #     breakpoint()
+    def __getattr__(cls: 'ObsPlusModel', item):
+        if item in getattr(cls, '__fields__', {}):
+            cls_name = cls.__name__
+            op_str = f"{cls_name}.{item}"
+            return OperationTracker(op_str)
+
+        msg = f"{cls.__name__} has not attribute {item}"
+        raise AttributeError(msg)
+
+
+class ObsPlusModel(BaseModel, metaclass=ObsPlusMeta):
+    # extra: Optional[Dict[str, Any]] = None
+    _contains = None  # for storing the containing info.
+
+    class Config:
+        pass
+        validate_assignment = True
+        arbitrary_types_allowed = True
+        orm_mode = True
+        extra = "allow"
+
+    @staticmethod
+    def _convert_to_obspy(value):
+        """Convert an object to obspy or return value. """
+        if hasattr(value, "to_obspy"):
+            return value.to_obspy()
+        return value
+
+    def to_obspy(self):
+        """Convert to obspy objects."""
+        name = self.__class__.__name__
+        name = 'ResourceIdentifier' if 'ResourceIdentifier' in name else name
+        cls = getattr(ev, name)
+        out = {}
+        # get schema and properties
+        schema = self.schema()
+        props = schema["properties"]
+        array_props = {x for x, y in props.items() if y.get("type") == "array"}
+        # iterate each property and convert back to obspy
+        for prop in props:
+            val = getattr(self, prop)
+            if prop in array_props:
+                out[prop] = [self._convert_to_obspy(x) for x in val]
+            else:
+                out[prop] = self._convert_to_obspy(val)
+        return cls(**out)
+
+    @classmethod
+    @lru_cache()
+    def get_graph_dict(cls) -> dict:
+        """return a mapper for this model."""
+        return _get_graph_dict(cls.schema())
+
+
+class ResourceIdentifier(ObsPlusModel):
+    """Resource ID"""
+
+    id: Optional[str] = None
+    _points_to = None
+
+    @root_validator(pre=True)
+    def get_id(cls, values):
+        """Get the id string from the resource id"""
+        value = values.get("id")
+        if value is None:
+            value = str(uuid4())
+        return {"id": value}
+
+    def __init_subclass__(cls, **kwargs):
+        super().__init_subclass__()
+        # set the name of what this points to (if specified)
[…]
