+++ obsplus/events/pd.py
@@ -214,7 +214,7 @@ def _get_eve_creation_info(event):
     return {x: out.get(x) for x in keys}
 
 
-@events_to_df.extractor()
+@events_to_df.extractor
 def _get_update_time(eve):
     """return the most recent time anything was updated in event"""
     creations = get_instances_from_tree(eve, cls=ev.CreationInfo)
