+++ tests/test_structures/test_model.py
@@ -1,16 +1,12 @@
 """
 Tests for the ObsPlusModel.
 """
-import pickle
 
-import glom
-import obspy.core.event as ev
 import pytest
-
-import obsplus
+import obsplus.events.schema as eschema
 from obsplus.exceptions import InvalidModelOperation, InvalidModelAttribute
 from obsplus.events.schema import Catalog, Event, Origin
-from obsplus.utils.model import ObsPlusModel, OperationTracker, ResourceIdentifier
+from obsplus.structures.model import ObsPlusModel, OperationTracker
 from obsplus.utils.misc import register_func
 
 
@@ -23,7 +19,6 @@ def catalog_graph_dict():
 class TestGraphDict:
     """Tests for getting the graph dictionary which defines relationships."""
 
-
     def test_expected_graph_keys(self, catalog_graph_dict):
         """ Ensure each of the models have a place in the graph. """
         for name in dir(eschema):
@@ -35,36 +30,37 @@ class TestGraphDict:
         Ensure the attr_id_ref, which tells us what kind of object the
         resource_id points to, is populated correctly on event.
         """
-        event_dict = catalog_graph_dict['Event']
-        ref_id_dict = event_dict['attr_id_ref']
+        event_dict = catalog_graph_dict["Event"]
+        ref_id_dict = event_dict["attr_id_ref"]
         assert len(ref_id_dict)
-        assert ref_id_dict['preferred_origin_id'] == 'Origin'
-        assert ref_id_dict['preferred_magnitude_id'] == 'Magnitude'
+        assert ref_id_dict["preferred_origin_id"] == "Origin"
+        assert ref_id_dict["preferred_magnitude_id"] == "Magnitude"
 
 
 class TestOperationTrackerBasic:
     """Tests for model attribute proxies."""
+
     proxy_list = []
 
-    @pytest.fixture(scope='class')
+    @pytest.fixture(scope="class")
     @register_func(proxy_list)
     def reference_id_proxy(self):
         """Return a proxy from a reference id"""
         return Event.preferred_origin_id
 
-    @pytest.fixture(scope='class')
+    @pytest.fixture(scope="class")
     @register_func(proxy_list)
     def list_id_proxy(self):
         """Return a proxy from a reference id"""
         return Catalog.events.preferred_origin_id
 
-    @pytest.fixture(scope='class')
+    @pytest.fixture(scope="class")
     @register_func(proxy_list)
     def time_id_proxy(self):
         """Return a proxy from a reference id"""
         return Origin.time
 
-    @pytest.fixture(scope='class')
+    @pytest.fixture(scope="class")
     @register_func(proxy_list)
     def chase_reference_id_proxy(self):
         """
@@ -73,7 +69,7 @@ class TestOperationTrackerBasic:
         """
         return Event.preferred_origin_id.time
 
-    @pytest.fixture(scope='class', params=proxy_list)
+    @pytest.fixture(scope="class", params=proxy_list)
     def model_proxy(self, request):
         """meta-fixture to aggregate proxies."""
         return request.getfixturevalue(request.param)
@@ -90,9 +86,9 @@ class TestOperationTrackerBasic:
     def test_or(self, list_id_proxy, reference_id_proxy):
         """Ensure or operator works"""
         out = list_id_proxy | reference_id_proxy
-        assert ' | ' in str(out)
+        assert " | " in str(out)
         # a second or is not currently allowed
-        with pytest.raises(InvalidModelOperation, match='Only one'):
+        with pytest.raises(InvalidModelOperation, match="Only one"):
             out | out
 
     def test_get_item(self, reference_id_proxy):
@@ -102,8 +98,8 @@ class TestOperationTrackerBasic:
 
[…]
