+++ obsplus/structures/model.py
@@ -39,8 +39,8 @@ class ObsPlusMeta(ModelMetaclass):
     #
     # def __setstate__(cls):
     #     breakpoint()
-    def __getattr__(cls: 'ObsPlusModel', item):
-        if item in getattr(cls, '__fields__', {}):
+    def __getattr__(cls: "ObsPlusModel", item):
+        if item in getattr(cls, "__fields__", {}):
             cls_name = cls.__name__
             op_str = f"{cls_name}.{item}"
             return OperationTracker(op_str)
@@ -70,7 +70,7 @@ class ObsPlusModel(BaseModel, metaclass=ObsPlusMeta):
     def to_obspy(self):
         """Convert to obspy objects."""
         name = self.__class__.__name__
-        name = 'ResourceIdentifier' if 'ResourceIdentifier' in name else name
+        name = "ResourceIdentifier" if "ResourceIdentifier" in name else name
         cls = getattr(ev, name)
         out = {}
         # get schema and properties
@@ -110,7 +110,7 @@ class ResourceIdentifier(ObsPlusModel):
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__()
         # set the name of what this points to (if specified)
-        name = cls.__name__.replace('ResourceIdentifier', '')
+        name = cls.__name__.replace("ResourceIdentifier", "")
         cls._points_to = name or cls._points_to
 
 
@@ -127,14 +127,12 @@ class _ModelWithResourceID(ObsPlusModel):
         return value
 
 
-
-
 class OperationTracker:
     """
     A simple class for keeping track of operations on a model.
     """
 
-    def __init__(self, op_str: Union[str, 'OperationTracker']):
+    def __init__(self, op_str: Union[str, "OperationTracker"]):
         if isinstance(op_str, self.__class__):
             self._op_str = op_str._op_str
         self._op_str = op_str
@@ -149,7 +147,7 @@ class OperationTracker:
         return f"AttributeTracker:: {self._op_str}"
 
     def __or__(self, other):
-        if ' | ' in self._op_str:
+        if " | " in self._op_str:
             msg = "Only one '|' operation is currently permitted."
             raise InvalidModelOperation(msg)
         new_str = f"({self._op_str} | {other._op_str})"
@@ -157,8 +155,10 @@ class OperationTracker:
 
     def __getitem__(self, item):
         if not isinstance(item, int):
-            msg = ('getitem is reserved for selecting elements from as list ' 
-                   f'and therefore must be an int. You passed {item}')
+            msg = (
+                "getitem is reserved for selecting elements from as list "
+                f"and therefore must be an int. You passed {item}"
+            )
             raise TypeError(msg)
         return self.__class__(f"{self._op_str}[{item}]")
 
@@ -182,10 +182,10 @@ class OperationTracker:
 
         # first resolve | by slicing out both strings
         op_str = self._op_str
-        if ' | ' in op_str:
-            paren_index1 = op_str.index('(')
-            paren_index2 = op_str.index(')')
-            inside_or_str = op_str[paren_index1 + 1: paren_index2]
+        if " | " in op_str:
+            paren_index1 = op_str.index("(")
+            paren_index2 = op_str.index(")")
+            inside_or_str = op_str[paren_index1 + 1 : paren_index2]
             resulting_type = self._validate_or(inside_or_str, graph_dict)
             op_str = op_str.replace(f"({inside_or_str})", resulting_type)
 
@@ -196,7 +196,7 @@ class OperationTracker:
 
     def _validate_or(self, or_str, graph_dict):
         """"""
-        str_1, str_2 = or_str.split(' | ')
+        str_1, str_2 = or_str.split(" | ")
         # first make sure attributes exist and are valid.
         obj1 = OperationTracker(str_1).validate(graph_dict)
[…]

[…]
