### 2021-02-04 ca4958f
fix seed_string (#219)
