### 2021-02-05 7039b61
updated bingham dataset, fix other small issues
