### 2021-02-05 61c4db0
updated bingham dataset, fix other small issues (#220)
