### 2021-02-23 5e72143
Deal with numpy deprecation about np.int
