### 2021-02-23 9e442dd
More numpy deprecation stuff
