### 2021-02-24 231acad
fix more numpy deprecation warnings
