### 2021-02-26 aa6dcb9
np.int Deprecation (#226)
* Deal with numpy deprecation about np.int

* More numpy deprecation stuff

* black

* fix more numpy deprecation warnings
