### 2021-03-10 f2d7cd1
refactor get_waveforms_bulk
