+++ tests/test_utils/test_waveforms_utils.py
@@ -2,15 +2,18 @@
 Tests for waveform utilities.
 """
 import copy
+
 import inspect
 from pathlib import Path
 
 import numpy as np
 import obspy
+import pandas as pd
 import pytest
 
 import obsplus
-from obsplus.constants import NSLC
+from obsplus.constants import NSLC, WAVEFORM_REQUEST_DTYPES
+from obsplus.exceptions import ValidationError
 from obsplus.interfaces import WaveformClient
 from obsplus.utils.time import to_timedelta64
 from obsplus.utils.waveforms import (
@@ -20,6 +23,7 @@ from obsplus.utils.waveforms import (
     merge_traces,
     stream_bulk_split,
     get_waveform_client,
+    get_waveform_bulk_df,
 )
 from obsplus.utils.testing import assert_streams_almost_equal
 
@@ -457,3 +461,67 @@ class TestStreamBulkSplit:
         assert len(out) == 1
         # without fill value this would only be 10 sec long
         assert abs(abs(out[0].stats.endtime - out[0].stats.starttime) - 20) < 0.1
+
+
+class TestGetWaveformBulk:
+    """Tests for getting the waveform bulk dataframe."""
+
+    times = (
+        "2010-01-01",
+        np.datetime64("2010-01-02"),
+        obspy.UTCDateTime("2010-11-20"),
+        obspy.UTCDateTime("2010-12-20").timestamp,
+    )
+
+    bulk1 = [
+        ("UU", "TMU", "01", "HHZ", times[0], times[1]),
+        ("UU", "NOQ", "01", "ENZ", times[2], times[3]),
+    ]
+
+    @pytest.fixture()
+    def bulk_df(self):
+        """Create a dataframe from bulk."""
+        df = pd.DataFrame(self.bulk1, columns=list(WAVEFORM_REQUEST_DTYPES))
+        out = get_waveform_bulk_df(df)
+        return out
+
+    def test_tuple(self):
+        """Ensure standard tuples produce bulk df."""
+        out = get_waveform_bulk_df(self.bulk1)
+        assert isinstance(out, pd.DataFrame)
+        assert len(out) == len(self.bulk1)
+
+    def test_dict(self):
+        """Ensure a list of dicts also works."""
+        bulk_dict = []
+        for bulk in self.bulk1:
+            req_dict = {i: v for i, v in zip(WAVEFORM_REQUEST_DTYPES, bulk)}
+            bulk_dict.append(req_dict)
+        df = pd.DataFrame(bulk_dict)
+        assert isinstance(df, pd.DataFrame)
+        assert len(df) == len(self.bulk1)
+
+    def test_dataframe(self, bulk_df):
+        """Ensure a datframe with no extra columns works."""
+        out = get_waveform_bulk_df(bulk_df)
+        assert isinstance(out, pd.DataFrame)
+        assert len(out) == len(self.bulk1) == len(bulk_df)
+
+    def test_dataframe_extra_column(self, bulk_df):
+        """The dataframe should work even with out of order/extra columns."""
+        df = bulk_df.copy()
+        df["bob"] = "lightening"
+        # reverse column order
+        df = df[reversed(list(df.columns))]
+        out = get_waveform_bulk_df(df)
+        assert isinstance(out, pd.DataFrame)
+        assert len(out) == len(self.bulk1) == len(bulk_df)
+        # the new columns should have been dropped
+        cols = list(WAVEFORM_REQUEST_DTYPES)
+        assert list(out.columns) == cols
+
+    def test_missing_column_raises(self, bulk_df):
+        """A missing column should raise."""
+        df = bulk_df.drop(columns=["network"])
+        with pytest.raises(ValidationError, match="network"):
+            get_waveform_bulk_df(df)
