### 2021-03-10 fdcac31
refactor filter function
