+++ obsplus/utils/waveforms.py
@@ -16,6 +16,7 @@ from obspy import Stream, UTCDateTime
 import obsplus
 from obsplus.constants import (
     waveform_clientable_type,
+    wave_type,
     NSLC,
     trace_sequence,
     NUMPY_FLOAT_TYPES,
@@ -27,7 +28,8 @@ from obsplus.exceptions import ValidationError
 from obsplus.interfaces import WaveformClient
 from obsplus.utils.pd import filter_index
 from obsplus.utils.pd import get_seed_id_series, cast_dtypes, _column_contains
-from obsplus.utils.time import to_utc
+from obsplus.utils.time import to_utc, to_timedelta64, to_datetime64
+from obsplus.utils.misc import ObjectWrapper
 
 
 # ---------- trim functions
@@ -121,25 +123,30 @@ def _trim_stream(df, stream, required_len, trim_tolerance):
     return stream.trim(starttime=t1, endtime=t2)
 
 
-def _stream_data_to_df(stream: Stream) -> pd.DataFrame:
+def _get_waveform_df(stream: wave_type) -> pd.DataFrame:
     """
-    Collect queryable params from stream, return dataframe.
+    Convert a stream of sequence of traces into a datframe.
 
     Parameters
     ----------
     stream
+        The streams to index
 
-    Returns
-    -------
+    Notes
+    -----
+    This is private because it is probably not quite polished enough to include
+    in the public API. More thought is needed how to do this properly.
     """
-    dtypes = {"starttime": "datetime64[ns]", "endtime": "datetime64[ns]"}
-    st_contents = [
-        tr.id.split(".") + [tr.stats.starttime._ns, tr.stats.endtime._ns]
-        for tr in stream
-    ]
-    columns = ["network", "station", "location", "channel", "starttime", "endtime"]
-    df = pd.DataFrame(st_contents, columns=columns)
-    return df.astype(dtypes)
+    stats_columns = list(NSLC) + ["starttime", "endtime", "sampling_rate"]
+    trace_contents = [{i: tr.stats[i] for i in stats_columns} for tr in stream]
+    df = pd.DataFrame(trace_contents, columns=stats_columns)
+    # ensure time(y) columns have proper
+    df["starttime"] = to_datetime64(df["starttime"])
+    df["endtime"] = to_datetime64(df["endtime"])
+    df["sampling_period"] = to_timedelta64(1 / df["sampling_rate"])
+    df["seed_id"] = get_seed_id_series(df)
+    df["trace"] = [ObjectWrapper(tr) for tr in stream]
+    return df
 
 
 def _get_bulk(bulk):
@@ -184,7 +191,7 @@ def stream_bulk_split(
         return []
 
     # # get dataframe of stream contents
-    sdf = _stream_data_to_df(st)
+    sdf = _get_waveform_df(st)
     # iterate stream, return output
     out = []
     for barg in bulk:
@@ -229,55 +236,45 @@ def merge_traces(st: trace_sequence, inplace=False) -> obspy.Stream:
     def _make_trace_df(traces: trace_sequence) -> pd.DataFrame:
         """ Create a dataframe form a sequence of traces. """
         # create dataframe of traces and stats (use ns for all time values)
-        data = [
-            {
-                "trace": x,
-                "nslc": x.id,
-                "sr": np.int64(
-                    np.round(1.0 / x.stats.sampling_rate, 9) * 1_000_000_000
-                ),
-                "start": x.stats.starttime._ns,
-                "end": x.stats.endtime._ns,  # TODO switch to .ns for obspy 1.2
-            }
-            for x in traces
-        ]
-        sortby = ["nslc", "sr", "start", "end"]
-        df = pd.DataFrame(data).sort_values(sortby)
[…]
