### 2021-03-10 400d819
fix de-duplicate on ebank.get_waveforms_bulk
