### 2021-03-15 d770e9e
add test case for overlap with high sample rate
