### 2021-03-15 e14aa53
add coverage, delete dead method
