### 2021-03-16 4e9cb02
refactor get_waveforms_bulk (#229)
* refactor get_waveforms_bulk

* refactor filter function

* refactor merge

* fix de-duplicate on ebank.get_waveforms_bulk

* add test case for overlap with high sample rate

* add coverage, delete dead method
