+++ versions/v0.1.1+11.g8ad3eab/_sources/notebooks/datastructures/events_to_pandas.ipynb.txt
@@ -1,0 +1,1186 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Events to DataFrame\n",
+    "ObsPlus provides a way to extract useful summary information from ObsPy objects in order to create dataframes. This transformation is lossy but very useful when the full complexity of the `Catalog` object isn't warranted. By default the `events_to_df` function collects information the authors of ObsPlus have found useful, but it is fully extensible/customizable using the functionality of the [DataFrameExtractor](../utils/dataframeextractor.ipynb)."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To demonstrate how the `Catalog` can be flattened into a table, let's again use the Crandall catalog."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-09-23T18:44:52.171980Z",
+     "iopub.status.busy": "2020-09-23T18:44:52.171065Z",
+     "iopub.status.idle": "2020-09-23T18:44:53.254702Z",
+     "shell.execute_reply": "2020-09-23T18:44:53.253663Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "import obspy\n",
+    "import numpy as np\n",
+    "from matplotlib import pyplot as plt\n",
+    "from obspy.clients.fdsn import Client\n",
+    "\n",
+    "import obsplus"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-09-23T18:44:53.265899Z",
+     "iopub.status.busy": "2020-09-23T18:44:53.264218Z",
+     "iopub.status.idle": "2020-09-23T18:44:55.010101Z",
+     "shell.execute_reply": "2020-09-23T18:44:55.009223Z"
+    }
+   },
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
+       "    .dataframe tbody tr th:only-of-type {\n",
+       "        vertical-align: middle;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe tbody tr th {\n",
+       "        vertical-align: top;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe thead th {\n",
+       "        text-align: right;\n",
+       "    }\n",
+       "</style>\n",
+       "<table border=\"1\" class=\"dataframe\">\n",
+       "  <thead>\n",
+       "    <tr style=\"text-align: right;\">\n",
+       "      <th></th>\n",
+       "      <th>time</th>\n",
+       "      <th>latitude</th>\n",
+       "      <th>longitude</th>\n",
+       "      <th>depth</th>\n",
+       "      <th>magnitude</th>\n",
+       "      <th>event_description</th>\n",
+       "      <th>associated_phase_count</th>\n",
+       "      <th>azimuthal_gap</th>\n",
+       "      <th>event_id</th>\n",
+       "      <th>horizontal_uncertainty</th>\n",
+       "      <th>...</th>\n",
+       "      <th>standard_error</th>\n",
+       "      <th>used_phase_count</th>\n",
+       "      <th>station_count</th>\n",
+       "      <th>vertical_uncertainty</th>\n",
[…]
