+++ versions/v0.1.0/_sources/notebooks/datasets/datasets.ipynb.txt
@@ -1,0 +1,611 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Datasets\n",
+    "\n",
+    "ObsPlus includes a few interesting datasets which are used for testing purposes, but a [template with instructions](https://github.com/niosh-mining/opsdata) is provided if you would like to create and distribute your own.\n",
+    "\n",
+    "\n",
+    "The datasets are \"lazy\" in that all but the most essential information will be downloaded only when some code requests the dataset. This helps keep the size of ObsPlus small, but does mean you will need an network connection the first time you use each dataset. Here are a few examples of things you can do with datasets:\n",
+    "\n",
+    "## Dataset basics\n",
+    "Loading a dataset only requires knowing its name (and having installed it, more on that later)."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "downloading waveform data for crandall_test dataset ...\n"
+     ]
+    },
+    {
+     "name": "stderr",
+     "output_type": "stream",
+     "text": [
+      "[2020-03-21 00:21:06,374] - obspy.clients.fdsn.mass_downloader - INFO: Initializing FDSN client(s) for http://service.iris.edu.\n",
+      "[2020-03-21 00:21:06,376] - obspy.clients.fdsn.mass_downloader - INFO: Successfully initialized 1 client(s): http://service.iris.edu.\n",
+      "[2020-03-21 00:21:06,376] - obspy.clients.fdsn.mass_downloader - INFO: Total acquired or preexisting stations: 0\n",
+      "[2020-03-21 00:21:06,377] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Requesting reliable availability.\n",
+      "[2020-03-21 00:21:06,789] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Successfully requested availability (0.41 seconds)\n",
+      "[2020-03-21 00:21:06,838] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Found 19 stations (57 channels).\n",
+      "[2020-03-21 00:21:06,843] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Will attempt to download data from 19 stations.\n",
+      "[2020-03-21 00:21:06,846] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Status for 57 time intervals/channels before downloading: NEEDS_DOWNLOADING\n",
+      "[2020-03-21 00:21:08,234] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Successfully downloaded 7 channels (of 7)\n",
+      "[2020-03-21 00:21:10,411] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Successfully downloaded 41 channels (of 50)\n",
+      "[2020-03-21 00:21:10,412] - obspy.clients.fdsn.mass_downloader - INFO: Client 'http://service.iris.edu' - Launching basic QC checks...\n",
[…]
