+++ versions/v0.2.0+1.g5ccea18/_sources/index.rst.txt
@@ -1,0 +1,26 @@
+.. image:: images/obsplus.png
+
+ObsPlus
+=======
+
+"A Pandas-Centric ObsPy Expansion Pack"
+
+`Introduction <notebooks/intro.ipynb>`_
+
+`Installation <notebooks/installation.ipynb>`_
+
+`Code <https://github.com/niosh-mining/obsplus>`_
+
+`Contributing <notebooks/contributing.ipynb>`_
+
+`Support Bug Reporting <notebooks/support.ipynb>`_
+
+`Quick Reference <quickref/index.rst>`_
+
+`Full API Documentation <api/obsplus.rst>`_
+
+:ref:`genindex`
+
+:ref:`modindex`
+
+:ref:`search`
