+++ versions/v0.1.1+18.g88c6ccb/_sources/api/obsplus.structures.dfextractor.rst.txt
@@ -1,0 +1,7 @@
+obsplus.structures.dfextractor module
+=====================================
+
+.. automodule:: obsplus.structures.dfextractor
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
