### 2021-03-20 69772e3
Cope with queries across the dateline
