### 2021-04-06 #2696 @calum-chamberlain
Sorry to pester,  but @leouieda are the any updates on this? I'm keen to have this out so that I can cite it for some upcoming papers for which ObsPlus had been very useful!
