### 2021-04-09 #2696 @ThomasLecocq
I can't check the boxes above, but I have no further remarks other than the citation for Pandas seem buggy in the latest rendered PDF. Other than that, it's good to go
