### 2021-04-09 #2696 @danielskatz
This should let you check the boxes
