### 2021-04-09 #2696 @ThomasLecocq
All 🟢 for me, just the missing Pandas citation in the PDF available above.
