+++ paper/paper.bib
@@ -3,8 +3,8 @@
   	Journal  = {Computational Science & Discovery},
   	Keywords = {seismology},
   	Title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
-    doi      = { 10.1088/1749-4699/8/1/014003 }
-  	Year     = 2015
+    doi      = { 10.1088/1749-4699/8/1/014003 },
+  	Year     = {2015},
 }
 
 
@@ -15,16 +15,16 @@
   	Keywords = {seismology},
   	Month    = mar,
   	Title    = {{QuakeML: status of the XML-based seismological data exchange format}},
-  	Year     = 2011
+  	Year     = {2011},
 }
 
 
-@InProceedings{ mckinney-proc-scipy-2010,
-  author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
-  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
-  pages     = { 56 - 61 },
-  year      = { 2010 },
-  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
-  doi       = { 10.25080/Majora-92bf1922-00a }
+@InProceedings{mckinney-proc-scipy-2010,
+  author    = {{W}es {M}c{K}inney},
+  title     = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+  booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+  pages     = {56 - 61},
+  year      = {2010},
+  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  doi       = {10.25080/Majora-92bf1922-00a},
 }
