### 2021-04-09 #2696 @d-chambers
Ok the pdf now compiles and I think the citation looks correct.
