### 2021-04-14 #2696 @kyleniemeyer
Hi @d-chambers, I'm the AEIC on duty this week, and stepping in here to help get this past the finish line. It looks like the reviewers have both recommended this for acceptance.

Some final steps needed:
- [ ] Your article is missing a Statement of Need section, which is a [required element](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of JOSS papers. You may be able to just move some existing text into that section.
- [ ] Once you've made that change, you can archive your software repository (e.g., on Zenodo) and provide the DOI here.
