### 2021-04-15 3d34a0a
Zenodo doi bage (#234)
* statement of need

* reword

* add badge

* move badge order
