### 2021-04-15 #2696 @kyleniemeyer
Thanks @d-chambers - can you clean up the author list on the current-version Zenodo record to match that of the paper?
