### 2021-04-15 #2696 @kyleniemeyer
@d-chambers sorry, looks like I missed something - the QuakeML is actually missing some key citation details, including a few authors, the volume (54) and number (1), and DOI: `10.4401/ag-4874`

Can you add these?
