### 2021-04-15 25f3f81
add details to bib
