+++ paper/paper.bib
@@ -1,30 +1,33 @@
 @article{Krischer:2015,
-  	Author   = {{Krischer}, L., {Megies}, T., {Barsch}, R., {Beyreuther}, M., {Lecocq}, T., {Caudron}, C. and {Wassermann}, J.,},
-  	Journal  = {Computational Science & Discovery},
-  	Keywords = {seismology},
-  	Title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
+  	author   = {{Krischer}, L., {Megies}, T., {Barsch}, R., {Beyreuther}, M., {Lecocq}, T., {Caudron}, C. and {Wassermann}, J.,},
+  	journal  = {Computational Science & Discovery},
+  	keywords = {seismology},
+  	title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
     doi      = { 10.1088/1749-4699/8/1/014003 },
-  	Year     = {2015},
+  	year     = {2015},
 }
-
 
 @article{Schorlemmer:2011,
-  	Author   = {{Schorlemmer}, D., {Euchner}, F., {Kästli}, P. and {Saul}, J.},
-  	Eprint   = {1703.04627},
-  	Journal  = {Annals of Geophysics},
-  	Keywords = {seismology},
-  	Month    = mar,
-  	Title    = {{QuakeML: status of the XML-based seismological data exchange format}},
-  	Year     = {2011},
+  	author   = {{Schorlemmer}, D., {Euchner}, F., {Kästli}, P. and {Saul}, J.},
+  	eprint   = {1703.04627},
+  	journal  = {Annals of Geophysics},
+  	keywords = {seismology},
+  	month    = mar,
+  	title    = {{QuakeML: status of the XML-based seismological data exchange format}},
+  	year     = {2011},
+  	volume   = {54},
+  	number   = {1},
+    doi      = {https://doi.org/10.3929/ethz-b-000034821},
+
 }
 
-
 @InProceedings{mckinney-proc-scipy-2010,
-  author    = {{W}es {M}c{K}inney},
-  title     = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
-  booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
-  pages     = {56 - 61},
+  author    = {{McKinney}, W.},
+  title     = {Data Structures for Statistical Computing in Python},
+  booktitle = {Proceedings of the 9th Python in Science Conference},
+  pages     = {56--61},
+  volume    = {445},
   year      = {2010},
-  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  editor    = {Stefan van der Walt and Jarrod Millman},
   doi       = {10.25080/Majora-92bf1922-00a},
 }
