### 2021-04-15 #2696 @d-chambers
Ok, I have added the volume, number and DOI.

The missing authors issues is interesting. The authors are included in the bib but don't show up in the PDF. Same issue for the ObsPy reference. I will dig into it more.
