### 2021-04-15 #2696 @kyleniemeyer
Ah, I think the issue related to the commas in the `author` fields.

For the ObsPy reference, it should be `{Krischer, L. and Megies, T. and Barsch, R. and Beyreuther, M. and Lecocq, T. and Caudron, C. and Wassermann, J.}` (you want "and" rather than commas between the authors)

Also, you also do not need brackets around all the surnames, except in uncommon cases.
