### 2021-04-15 #2696 @kyleniemeyer
and then `{Schorlemmer, D. and Euchner, F. and Kästli, P. and Saul, J.}` for QuakeML
