+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Krischer:2015,
-  	author   = {{Krischer}, L., {Megies}, T., {Barsch}, R., {Beyreuther}, M., {Lecocq}, T., {Caudron}, C. and {Wassermann}, J.,},
+  	author   = {{Krischer}, L. and {Megies}, T. and {Barsch}, R. and {Beyreuther}, M. and {Lecocq}, T. and {Caudron}, C. and {Wassermann}, J.},
   	journal  = {Computational Science & Discovery},
   	keywords = {seismology},
   	title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
@@ -8,7 +8,7 @@
 }
 
 @article{Schorlemmer:2011,
-  	author   = {{Schorlemmer}, D., {Euchner}, F., {Kästli}, P. and {Saul}, J.},
+  	author   = {{Schorlemmer}, D. and {Euchner}, F. and {Kästli}, P. and {Saul}, J.},
   	eprint   = {1703.04627},
   	journal  = {Annals of Geophysics},
   	keywords = {seismology},
