### 2021-04-15 #2696 @kyleniemeyer
@d-chambers ah, one final issue: please remove the `https://doi.org/` from the QuakeML DOI, it should just include the DOI itself.
