+++ paper/paper.bib
@@ -3,7 +3,7 @@
   	journal  = {Computational Science & Discovery},
   	keywords = {seismology},
   	title    = {{ObsPy: A bridge for seismology into the scientific Python ecosystem}},
-    doi      = { 10.1088/1749-4699/8/1/014003 },
+    doi      = {10.1088/1749-4699/8/1/014003},
   	year     = {2015},
 }
 
@@ -17,7 +17,7 @@
   	year     = {2011},
   	volume   = {54},
   	number   = {1},
-    doi      = {https://doi.org/10.3929/ethz-b-000034821},
+    doi      = {10.3929/ethz-b-000034821},
 
 }
 
